move DailyUserDigest
